refactor(telemetry): reduce to PR 1 — host event parameters only
SCAL-333657

Splitting this work up, and this branch keeps only the first slice: which host
event is triggered and which parameters it uses.

No existing SDK behaviour changes. `visual-sdk-trigger-<HostEvent>` keeps its
name, its timing and its call site; it just carries properties now, where
before it carried none. One existing file is touched, by ten lines.

Everything else is removed from this branch and preserved at bb33dd47 for the
follow-up PRs: the outcome and duration of a trigger, the response story for
both directions, embed event tracking, the registration-noise work, the
custom-handler timeout fix, and the pre-init queue cap.

Dropped outright, per review: `isRegisteredBySDK`. The flag was broken, and
propagating it meant `{ start: false }, true` at every internal call site to
reach a positional parameter past a default. Whether a host application
listens to an embed event is better answered by counting the handlers a
dispatch actually ran, which needs no flag and belongs with the embed event
work anyway.
