+++ src/embed/app.ts
@@ -1023,15 +1023,12 @@ export class AppEmbed extends V1Embed {
         viewConfig.embedComponentType = 'AppEmbed';
         super(domSelector, viewConfig);
         if (this.viewConfig.fullHeight === true) {
-            this.on(
-                EmbedEvent.RouteChange,
-                this.setIframeHeightForNonEmbedLiveboard, { start: false }, true,
-            );
-            this.on(EmbedEvent.EmbedHeight, this.updateIFrameHeight, { start: false }, true);
-            this.on(EmbedEvent.EmbedIframeCenter, this.embedIframeCenter, { start: false }, true);
+            this.on(EmbedEvent.RouteChange, this.setIframeHeightForNonEmbedLiveboard);
+            this.on(EmbedEvent.EmbedHeight, this.updateIFrameHeight);
+            this.on(EmbedEvent.EmbedIframeCenter, this.embedIframeCenter);
             this.on(
                 EmbedEvent.RequestVisibleEmbedCoordinates,
-                this.requestVisibleEmbedCoordinatesHandler, { start: false }, true,
+                this.requestVisibleEmbedCoordinatesHandler,
             );
         }
     }
