+++ src/utils/hostEventTelemetry.ts
